+++ src/bundle.rs
@@ -1,6 +1,9 @@
 //! Structs related to bundles of Orchard actions.
 
+mod batch;
 pub mod commitments;
+
+pub use batch::BatchValidator;
 
 use core::fmt;